chore: remove all code from reth-primitives, keep only re-exports
Remove all submodules (block, receipt, transaction), benchmarks, and
dev-dependencies. Inline the remaining re-exports directly in lib.rs.

Amp-Thread-ID: https://ampcode.com/threads/T-019c7ea1-f3e8-7438-ae05-3e3e1d2f54a4
