--- crates/primitives/src/receipt.rs
@@ -1,5 +1,0 @@
-/// Retrieves gas spent by transactions as a vector of tuples (transaction index, gas used).
-pub use reth_primitives_traits::receipt::gas_spent_by_transactions;
-
-/// Receipt containing result of transaction execution.
-pub use reth_ethereum_primitives::Receipt;
