--- crates/primitives/src/transaction/signature.rs
@@ -1,1 +1,0 @@
-pub use reth_primitives_traits::crypto::secp256k1::{recover_signer, recover_signer_unchecked};
