--- crates/primitives/src/transaction/util.rs
@@ -1,3 +1,0 @@
-//! Utility functions for signature.
-
-pub use reth_primitives_traits::crypto::*;
